taskaction delete permission control add

+++ src/main/java/com/yil/workflow/service/TaskActionService.java
@@ -1,13 +1,7 @@
 package com.yil.workflow.service;
 
-import com.yil.workflow.dto.TaskActionDocumentRequest;
-import com.yil.workflow.dto.TaskActionMessageRequest;
-import com.yil.workflow.dto.TaskActionRequest;
-import com.yil.workflow.dto.TaskActionResponse;
+import com.yil.workflow.dto.*;
 import com.yil.workflow.exception.*;
-import com.yil.workflow.model.Action;
-import com.yil.workflow.model.ActionSource;
-import com.yil.workflow.model.GroupUser;
 import com.yil.workflow.model.TaskAction;
 import com.yil.workflow.repository.TaskActionDao;
 import lombok.RequiredArgsConstructor;
@@ -17,7 +11,6 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.Date;
-import java.util.List;
 
 @RequiredArgsConstructor
 @Service
@@ -27,9 +20,7 @@ public class TaskActionService {
     private final ActionService actionService;
     private final TaskActionDocumentService taskActionDocumentService;
     private final TaskActionMessageService taskActionMessageService;
-    private final StepService stepService;
     private final ActionSourceService actionSourceService;
-    private final GroupUserService groupUserService;
 
     public static TaskActionResponse toDto(TaskAction taskAction) throws NullPointerException {
         if (taskAction == null)
@@ -42,73 +33,48 @@ public class TaskActionService {
     }
 
     @Transactional(rollbackFor = {Throwable.class})
-    public TaskActionResponse save(TaskActionRequest request, long taskId, long userId) throws ActionNotFoundException, NotAvailableActionException, YouDoNotHavePermissionException, StepNotFoundException {
+    public TaskActionResponse save(TaskActionRequest request, long taskId, long userId) throws ActionNotFoundException, NotAvailableActionException, YouDoNotHavePermissionException, StepNotFoundException, StartUpActionException, NotNextActionException {
         TaskAction currentTaskAction = getLastAction(taskId);
 
         //region permission control
         if (currentTaskAction == null) {
-            boolean state = false;
-            List<ActionSource> actionSources = actionSourceService.findAllByActionIdAndTargetTypeId(request.getActionId(), 3);
-            for (ActionSource actionSource : actionSources) {
-                if (groupUserService.existsById(new GroupUser.Pk(actionSource.getGroupId(), userId, 3))) {
-                    state = true;
-                    break;
-                }
-            }
-            if (!state)
+
+            if (!actionSourceService.userInActionGroup(request.getActionId(), userId))
                 throw new YouDoNotHavePermissionException();
         } else {
             TaskAction firstAction = null;
-            try {
-                if (currentTaskAction.getParentId() == null) //current action parent id is null then first action
-                    firstAction = findByTaskIdOrderByIdAsc(taskId);
-                else
-                    firstAction = currentTaskAction;
-            } catch (Exception e) {
-
-            }
+            if (currentTaskAction.getParentId() == null) //current action parent id is null then first action
+                firstAction = taskActionDao.findByTaskIdOrderByIdAsc(taskId).orElse(null);
+            else
+                firstAction = currentTaskAction;
             boolean state = false;
-            List<ActionSource> actionSources = actionSourceService.findAllByActionId(request.getActionId());
-            for (ActionSource actionSource : actionSources) {
-                if (actionSource.getTargetTypeId().equals(1)) { // task creator
-                    if (firstAction != null && firstAction.getCreatedUserId().equals(userId)) { // first task creator is user ?
-                        state = true;
-                        break;
-                    }
-                } else if (actionSource.getTargetTypeId().equals(2)) { //last action user
-                    if (currentTaskAction.getCreatedUserId().equals(userId)) // last action is user ?
-                    {
-                        state = true;
-                        break;
-                    }
-                } else if (actionSource.getTargetTypeId().equals(3)) {
-                    if (groupUserService.existsById(new GroupUser.Pk(actionSource.getGroupId(), userId, 3))) {
-                        state = true;
-                        break;
-                    }
-                }
+            if (firstAction != null &&
+                    firstAction.getCreatedUserId() != null &&
+                    actionSourceService.existsByActionIdAndTargetTypeId(request.getActionId(), TargetTypeService.TaskCreator)) {
+                state = true;
+            } else if (currentTaskAction.getCreatedUserId() != null &&
+                    actionSourceService.existsByActionIdAndTargetTypeId(request.getActionId(), TargetTypeService.LastActionUser)) {
+                state = true;
+            } else if (actionSourceService.userInActionGroup(request.getActionId(), userId)) {
+                state = true;
             }
             if (!state)
                 throw new YouDoNotHavePermissionException();
         }
         //endregion permission control
 
-        Action action;
-
         //region action control
-        if (currentTaskAction == null) { // yeni ise başlangıç aksiyonu mu ?
-            action = actionService.findByIdAndDeletedTimeIsNull(request.getActionId());
-            if (!stepService.existsByIdAndStepTypeIdAndEnabledTrueAndDeletedTimeIsNull(action.getStepId(), 1)) //action stepi aktif ve başlangıç step mi ?
-                throw new StepNotFoundException();
+        if (currentTaskAction == null && !actionService.isStartUpAction(request.getActionId())) { // yeni ise başlangıç aksiyonu mu ?
+            throw new StartUpActionException();
         } else {
-            Action currentAction = actionService.findByIdAndDeletedTimeIsNull(currentTaskAction.getActionId()); //sonraki adım doğrumu ?
-            action = actionService.findByIdAndEnabledTrueAndDeletedTimeIsNull(request.getActionId(), currentAction.getNextStepId());
+            if (!actionService.isNextAction(currentTaskAction.getActionId(), request.getActionId()))
+                throw new NotNextActionException();
         }
         //endregion action control
 
         TaskAction taskAction = new TaskAction();
         taskAction.setTaskId(taskId);
-        taskAction.setActionId(action.getId());
+        taskAction.setActionId(request.getActionId());
         taskAction.setParentId(currentTaskAction != null ? currentTaskAction.getId() : null);
         taskAction.setCreatedUserId(userId);
         taskAction.setCreatedTime(new Date());
@@ -135,13 +101,17 @@ public class TaskActionService {
         return taskActionDao.getLastAction(taskId);
     }
 
-    @Transactional(readOnly = true)
-    public TaskAction findByTaskIdOrderByIdAsc(long taskId) throws TaskActionNotFoundException {
-        return taskActionDao.findByTaskIdOrderByIdAsc(taskId).orElseThrow(() -> new TaskActionNotFoundException());
-    }
-
+    /**
+     * Sadece aksiyonu icra eden silebilir
+     *
+     * @param taskActionId
+     * @param userId
+     * @throws YouDoNotHavePermissionException
+     */
     @Transactional(rollbackFor = {Throwable.class})
     public void delete(long taskActionId, long userId) throws YouDoNotHavePermissionException {
+        if (!taskActionDao.existsByIdAndCreatedUserId(taskActionId, userId))
+            throw new YouDoNotHavePermissionException();
         taskActionDao.deleteById(taskActionId);
     }
